Refactor phase 2c: split gpio.py into a package
Continues the effort tracked in docs/development/LARGE_FILE_REFACTOR_PLAN.md
and completes phase 2. Pure motion - no behaviour was altered.

app_utils/gpio.py (3149 lines) became the app_utils/gpio/ package. Four
independent subsystems shared one file: the backend abstraction (lgpio,
sysfs and null behind one Protocol, plus gpiozero pin-factory setup), the
GPIOController and its behaviour manager, the NeoPixel strip controller,
and the USB tower-light controller. New layout: pin_types (166),
backends (426), tower_light (410), neopixel (336), controller (1003),
behavior (546) and config_loaders (440), with an acyclic dependency
graph. Each optional-dependency probe now sits with its only consumer -
get_gpio_settings/_GPIO_SETTINGS_AVAILABLE with the database loaders,
PixelStrip/NeopixelColor/_NEOPIXEL_LIB_AVAILABLE with the NeoPixel
controller. The package __init__ re-exports all 72 names the single-file
module exposed, so all six production importers are unchanged; each was
imported to confirm.

Verified as pure motion by comparing ast.dump() of every top-level
definition: 28 definitions, 28 matches, zero differences, plus an
assertion that every non-blank line of the original landed in exactly
one module.

Retarget 31 monkeypatch.setattr sites to the module that uses each name
(gpio.config_loaders, gpio.neopixel, gpio.controller, gpio.behavior.time).
Rebinding on the re-exporting package does not change what those modules
resolve from their own globals. No assertion changed. Confirmed
load-bearing: pointing the patches back at the package fails 9 tests.

Two lessons, both now in the plan's new pre-split checklist. The
enums-and-dataclasses module was first called types.py, which shadows the
stdlib types module for any process whose cwd is the package directory -
the stdlib chain dataclasses -> re -> enum -> types then fails with a
partially-initialised-module error. It is pin_types.py, which is also
more descriptive. And grep found only 24 of the 31 patch sites; it missed
multi-line setattr calls and a bare gpio.time.sleep attribute read, so
the checklist now calls for an AST scan of the tests instead.

Tests: full suite 1973 passed, 67 xfailed, 0 failures.
